+++ Medidata.Cloud.ExcelLoader/ColumnDefinition.cs
@@ -4,11 +4,6 @@ namespace Medidata.Cloud.ExcelLoader
 {
     public class ColumnDefinition : IColumnDefinition
     {
-        public static dynamic New
-        {
-            get { return new ColumnBuilder(); }
-        }
-
         public Type PropertyType { get; set; }
         public string PropertyName { get; set; }
         public string Header { get; set; }
